SCRUM-24 Establish Core API Layer with tRPC
SCRUM-24 CHORE Establish Core API Layer with tRPC

- install dependencies for tRPC packages
- set up init tRPC
- set up root App Router exposed via Nest.js API route with a sample route to test/refer. (server-side tRPC router)
- set up tRPC client and React Query provider on front end

As of now, tRPC is successful when using local dummy data. tRPC call is successful and data is shown on home page.

Will commit here to safe and udpate later with tRPC context when prisma is set up.

Rebasing to update prisma client and fix for persistent volume.#